requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-genai==0.2.2
python-docx==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1
